Rename TestParquetPageSkipping to TestHiveTestParquetPageSkipping
Additionally, extract a method to return table definition
to allow running in Iceberg connector.
