+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestHiveTestParquetPageSkipping.java
@@ -1,0 +1,62 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.trino.plugin.hive;
+
+import com.google.common.collect.ImmutableMap;
+import io.trino.testing.QueryRunner;
+
+import static java.lang.String.format;
+
+public class TestHiveTestParquetPageSkipping
+        extends BaseTestParquetPageSkipping
+{
+    @Override
+    protected QueryRunner createQueryRunner()
+            throws Exception
+    {
+        return HiveQueryRunner.builder()
+                .setHiveProperties(
+                        ImmutableMap.of(
+                                "parquet.use-column-index", "true",
+                                // Small max-buffer-size allows testing mix of small and large ranges in HdfsParquetDataSource#planRead
+                                "parquet.max-buffer-size", "400B"))
+                .build();
+    }
+
+    @Override
+    protected String tableDefinitionForSortedTables(String tableName, String sortByColumnName, String sortByColumnType)
+    {
+        String createTableTemplate =
+                "CREATE TABLE %s ( " +
+                        "   orderkey bigint, " +
+                        "   custkey bigint, " +
+                        "   orderstatus varchar(1), " +
+                        "   totalprice double, " +
+                        "   orderdate date, " +
+                        "   orderpriority varchar(15), " +
+                        "   clerk varchar(15), " +
+                        "   shippriority integer, " +
+                        "   comment varchar(79), " +
+                        "   rvalues double array " +
+                        ") " +
+                        "WITH ( " +
+                        "   format = 'PARQUET', " +
+                        "   bucketed_by = array['orderstatus'], " +
+                        "   bucket_count = 1, " +
+                        "   sorted_by = array['%s'] " +
+                        ")";
+        createTableTemplate = createTableTemplate.replaceFirst(sortByColumnName + "[ ]+([^,]*)", sortByColumnName + " " + sortByColumnType);
+        return format(createTableTemplate, tableName, sortByColumnName);
+    }
+}
